rename xgb module to xgboost
