+++ src/tabular_trees/__init__.py
@@ -1,6 +1,6 @@
 """tabular_trees"""
 
 from ._version import __version__
-from . import xgb
+from . import xgboost
 from . import helpers
 from . import trees
