+++ tests/xgb/test_parser.py
@@ -22,16 +22,16 @@ def test_text_json_parsing_equal(tmp_path):
     model.dump_model(json_dump_no_stats, with_stats=False, dump_format="json")
 
     # parse dumped files
-    tree_df1 = tabular_trees.xgb.parser._read_dump_text(
+    tree_df1 = tabular_trees.xgboost.parser._read_dump_text(
         text_dump, return_raw_lines=False
     )
-    tree_df2 = tabular_trees.xgb.parser._read_dump_text(
+    tree_df2 = tabular_trees.xgboost.parser._read_dump_text(
         text_dump_no_stats, return_raw_lines=False
     )
-    tree_df3 = tabular_trees.xgb.parser._read_dump_json(
+    tree_df3 = tabular_trees.xgboost.parser._read_dump_json(
         json_dump, return_raw_lines=False
     )
-    tree_df4 = tabular_trees.xgb.parser._read_dump_json(
+    tree_df4 = tabular_trees.xgboost.parser._read_dump_json(
         json_dump_no_stats, return_raw_lines=False
     )
