+++ tests/xgb/test_shap_equality.py
@@ -23,9 +23,9 @@ def test_prediction_decomposition_eqal_eli5():
 
     row_data = data.iloc[0]
 
-    tree_df = tabular_trees.xgb.parser.parse_model(model)
+    tree_df = tabular_trees.xgboost.parser.parse_model(model)
 
-    shapley_values = tabular_trees.xgb.explainer.shapley_values(
+    shapley_values = tabular_trees.xgboost.explainer.shapley_values(
         tree_df.tree_data, row_data, False
     )
